requirements.txt
aniso8601==9.0.1
asgiref==3.4.1
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-RESTful==0.3.9
h11==0.12.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pymongo==4.0.1
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
uvicorn==0.16.0
Werkzeug==2.0.2
